Add Sun/NetBSD hostapi called audioio (#645)
* Add Sun/NetBSD hostapi

This hostapi has been available in the pkgsrc package for PortAudio
for over a year (we build for Solaris and NetBSD). It was inspired by
problems interfacing between PortAudio's OSS hostapi and NetBSD's OSS
compatibility layer that I couldn't fix without adding very ugly hacks
to the OSS hostapi - in the end I decided to write a new, clean one for
the native interface.

Tested building with the autotools and CMake build systems.
Test and example programs work.