cmake: Assign library .so version for Linux statically as 2. (#686)
I am no library versioning expert, but using ${PROJECT_VERSION} as
SOVERSION in the CMakeLists.txt file builds a dynamic libportaudio.so
with a SOVERSION (cfe. `objdump -p libportaudio.so | grep SONAME`) of
libportaudio.so.19.8.

This means that any executable compiled/linked against libportaudio.so
will encode a runtime dependency on libportaudio.so.19.8, whereas for
older v19.7 portaudio and earlier, the dependency was on libportaudio.so.2.

That means that any future application binary will have a dependency on
the specific portaudio version it was built against, even though it would
be api and abi compatible to other or older versions of libportaudio, so
upgrades of libportaudio.so to other versions become impossible without
rebuilding all applications that use libportaudio on that machine, or
playing tedious hacky games with symbolic links.

Hardcode the SOVERSION to 2, so the future libportaudio v19.8 will stay
compatible with applications built against v19.7 and earlier. I am not
sure if the old autotools/automake style build scripts also need a similar
update, as i only use the new cmake system.

My understanding is that the soversion should only be incremented if
libportaudio receives backwards incompatible abi changes, not on each
minor version release.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>