alsa: Fix distorted sound from half-duplex capture. (#657)
If an audio stream is only opened for capture in half-duplex
mode, then the self->playback.canMmap variable is false by
default. This triggers setup of a hostBufferSizeMode of
paUtilBoundedHostBufferSize, something that is only meant to
happen for playback streams under some conditions. This leads
to unnecessary processing.

Fix this by also checking for self->playback.pcm, ie. that
the playback component is actually used, so this setting
only gets applied for half-duplex playback or full-duplex
playback + capture streams if needed.

Tested to work correctly now for playback, capture, and
full-duplex playback + capture.

The specific commit which introduced the regression in May 2009 is this one:
https://github.com/PortAudio/portaudio/commit/85a522fc4abb130b8e15131b1345bb857505c333

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>