Update compile_windows_mingw-w64.dox
The sawtooth example is absurdly loud. If it's left as part of the default build steps, it will probably cause some hearing damage. This moves it into a separate section with a warning.

An alternative fix would be to lower the sawtooth example to a reasonable volume.

Also, I added a note about msys2's default package.