dsound: use paUtilBoundedHostBufferSize
Currently the DirectSound host API uses the buffer processor in the
"paUtilVariableHostBufferSizePartialUsageAllowed" buffer size mode.
What's special about this mode is that it allows the buffer processor to
leave input frames in the host buffer if there are not enough frames
available to fire the stream callback. In practice the buffer processor
will only do this in full duplex mode.

However, just because we *can* leave these frames in the host buffer,
doesn't mean we *should*.

It is not clear to me why the DS code was set up to leave data in the
host input buffer. The reason stated in a code comment, "because DS can
split the host buffer when it wraps around" doesn't make sense to me -
the buffer processor seems perfectly capable of handling buffer split
and wraparound (through functions such as
PaUtil_Set2ndInputFrameCount()) regardless of the buffer size mode.

While it is not clear what could be the upside of leaving data in the
host buffer, the downsides seem evident: the longer data stays in the
input buffer, the more likely it is to get trampled on by the capture
cursor (buffer overflow). Copying the data out of the host input buffer
and into the buffer processor temp buffer at the earliest opportunity
seems like a safer option with no downside.

Another downside of this mode is that currently, the behavior with
respect to the host *output* buffer seems... confused, with the buffer
processor writing a number of frames to the output host buffer that is
inconsistent with the number of "frames processed" it returns to the DS
code, leading to #770. Arguably we could try to fix the buffer
processor, but why fix a mode whose reason to exist is unclear in the
first place?

Another oddity of this mode is that it leads to the DS code locking the
DS input and output buffers for read/write, but the buffer processor
might then decide to do nothing with them. These DS Lock/Unlock calls
are therefore wasted, reducing efficiency.

This commit changes the mode to paUtilBoundedHostBufferSize. Not only
does this fix #770, it also has the nice side effect of having the
buffer processor provision a temp buffer to match the host buffer size
(if paFramesPerBufferUnspecified is used), avoiding unnecessary user
buffer splitting.

This change was tested using paloopback to test for glitches with
various framesPerBuffer and suggested latency parameters, both in half
duplex and full duplex mode. No regressions were observed, and
paloopback results show that #770 is fixed by this change.

Fixes #770