paloopback: fix wrong test duration
This fixes a regression introduced in 5b5feb3 where, during sample rate
testing, the test duration would be computed based on the global sample
rate instead of the chosen sample rate, resulting in the test running
for too long or not long enough depending on the sample rate being
tested. The main symptom is that paloopback would complain about
timeouts when testing sample rates 8000, 11000 and 16000.