dsound: simplify CalculateBufferSettings()
This commit rewrites CalculateBufferSettings() in an attempt to make it
a bit less headache-inducing. The basic idea is to reduce branching,
reduce code duplication, and share code paths between the various cases
(fixed/variable user buffer, half/full duplex) as much as possible.
Also, min()/max() are easier to read than if statements.

This is a pure refactoring - there should be no change in observable
behaviour. The math stays the same, it is merely reshuffled around in
place.