dsound: clamp smaller buffer sizes to 62.5 ms on Vista and later
This works around a DirectSound limitation where input host buffer sizes
smaller than 31.25 ms are basically unworkable and make PortAudio hang.
The workaround is to impose a minimal buffer size of 2*31.25 ms on
input-only and full-duplex streams. This is enough for the read cursor
to advance twice around the buffer, basically resulting in de facto
double buffering.

This change was tested with `paloopback` under a wide variety of
half/full-duplex, framesPerBuffer, and suggested latency parameters.
(Note the testing was done on top of #772 as otherwise paloopback is not
usable.)

NOTE: in the input+output case this patch also increases the output
buffer size.

Fixes #775