Treat warnings as errors on CMake CI (#780)
* Add a CMake option to turn warnings into errors

The option is disabled by default as it can get in the way of people
merely building PortAudio (as opposed to doing development work). This
is provided as a CMake option so that it can easily be switched on by
people doing local development. (The option is turned on in CI.)

In this mode pre-existing warnings are suppressed, in other words they
are "grandfathered" in for the sake of preventing build breakage. The
idea is to enable enforcement first to prevent backsliding, and then we
can take care of existing warnings later.

* Enforce no warnings through the CMake CI

This will prevent backsliding by failing CI on any code that produces
new types of warnings.