Pulseaudio: Handle Mono to stereo
Honestly I'm no sure why this works!
I probably should ask from ChatGPT but this
commit makes Mono to Stereo work as expected.

If Input and Output are mono then left channel
is used which should be correct thing to do.
if input is mono and output is stereo then
it's converted to stereo output from mono.

Please note! As I don't see usage for Input stereo
to mono output I did not make that work.
It should be fixed in some distant future if
needed. It does not work in other HostAPI's as
far as I read the code.. maybe AI would help