JACK: try to use stable port names
JACK connection managers save/load the connection graph so users don't
need to create connections manually every time. This depends on stable
port names but PortAudio currently generates incrementing port names
throughout the lifetime of a running process.

Reset the port name counters when all streams are closed so that the
same port names are generated when the streams are opened again.

This is not a perfect solution because applications that open streams in
different orders each time or never close all open streams before
restarting audio processing will not reset the port counters. However,
typical applications should work with this simple approach and it's an
improvement over current behavior.

Tested with Audacity 3.3.3 on Fedora 39.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>