Make PaStream a stronger type
This changes PaStream* from a pointer to void to a pointer to an
undefined, but unique type. This ensures functions that take a PaStream*
reject incompatible types, for improved type safety.

While this is technically an API change, it is unlikely to break anyone
as it's unlikely user code depends on this type being void specifically
(unless they suffer from the bugs that this change is designed to
prevent!).

Fixes #915