Make MME wait forever for the thread to exit
See #919 for rationale.

Note that the previous Pa_StopStream() uses a bounded timeout to abort
the stream if it doesn't stop within the alloted time, and then another
bounded timeout to wait for the streaming thread to exit. This commit
preserves the first timeout as it makes sense and appears to be safe,
but gets rid of the second one so that we wait indefinitely for the
streaming thread to exit.