Remove PA_WIN_DS_USE_WMME_TIMER
This build flag was introduced 13 years ago to allow PortAudio users to
roll back to some previous MME timer-based DirectSound implementation.

As far as I can tell no-one has ever used that flag (Google returns ~0
hits for it), and presumably no-one is going to bother testing with this
flag in place. Get rid of it to simplify the DS code which currently
looks like an #ifdef fest.