Wait for the DS thread when stopping
The previous code would return from Pa_StopStream() when the streaming
thread signals that it is done. This is racy, because the streaming
thread is technically still running code by that point. If, for example,
the PortAudio user decides to unload the entire PortAudio DLL as soon as
Pa_StopStream() returns, hilarity could ensue.

Instead we can simply wait on the thread handle itself, which is both
safer and simpler.

Fixes #926